Add new protection coordinator.
