+++ Trust/AppDelegate.swift
@@ -7,6 +7,7 @@ import Lokalise
 class AppDelegate: UIResponder, UIApplicationDelegate, UISplitViewControllerDelegate {
     var window: UIWindow?
     var coordinator: AppCoordinator!
+    var protectionCoordinator: ProtectionCoordinator!
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
         window = UIWindow(frame: UIScreen.main.bounds)
         do {
@@ -16,17 +17,19 @@ class AppDelegate: UIResponder, UIApplicationDelegate, UISplitViewControllerDele
         } catch {
             print("EtherKeystore init issue.")
         }
+        //This is separate coordinator for the protection of the sensitive information.
+        protectionCoordinator = ProtectionCoordinator()
         return true
     }
     func application(_ application: UIApplication, didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data) {
         coordinator.didRegisterForRemoteNotificationsWithDeviceToken(deviceToken: deviceToken)
     }
     func applicationWillResignActive(_ application: UIApplication) {
-        coordinator.applicationWillResignActive()
+        protectionCoordinator.applicationWillResignActive()
     }
     func applicationDidBecomeActive(_ application: UIApplication) {
         Lokalise.shared.checkForUpdates { _, _ in }
-        coordinator.applicationDidBecomeActive()
+        protectionCoordinator.applicationDidBecomeActive()
     }
     func application(_ application: UIApplication, shouldAllowExtensionPointIdentifier extensionPointIdentifier: UIApplicationExtensionPointIdentifier) -> Bool {
         if extensionPointIdentifier == UIApplicationExtensionPointIdentifier.keyboard {
